+++ lib/commands/index.d.ts
@@ -8,5 +8,13 @@ export declare const commands: (CommandEvent: CommandEvent) => Promise<
       disabled: number;
       failed: number;
     }
+  | {
+      total_reps_with_accounts_ids: number;
+      repzo_reps: number;
+      unset: number;
+      created: number;
+      updated: number;
+      failed: number;
+    }
 >;
 export declare const commandsList: Command[];