+++ lib/types.d.ts
@@ -73,5 +73,6 @@ export declare type FailedDocsReport = {
   doc_id?: string;
   doc?: any;
   error_message: any;
+  [key: string]: any;
 }[];
 export {};